Update prompts for empty responses and added prompt for Xcode download because of size

+++ environments/mobiledev.sh
@@ -10,7 +10,14 @@ brew install bluepill
 brew install detekt
 
 # Install Xcode
-mas install 497799835
+read -r -p 'Download Xcode (might be bandwidth hungry) ? (Y/n) : ' download_xcode
+if [ "$download_xcode" = "" ] || [ "$download_xcode" = "y" ] || [ "$download_xcode" = "Y" ] || [ "$download_xcode" = "yes" ]
+then
+  mas install --force 497799835
+else
+  mas install 497799835
+fi
+
 
 # Install Apple COnfigurator 2
 mas install 1037126344
